feat: add redux for Formpage
